Removido CombatProcesssor
